Add version data to game model
